refactor(test): extract build_pdf_with_content_stream to tests/common/
